Added a bottleneck library to control the api calls rate to gmail.
It wraps the api call method with a predefined limiter.
