+++ package.json
@@ -32,6 +32,7 @@
     "webpack": "^4.44.1",
     "webpack-cli": "^3.3.12",
     "webpack-dev-server": "^3.11.0",
-    "gapi-script": "^1.0.2"
+    "gapi-script": "^1.0.2",
+    "bottleneck": "~2.19.5"
   }
 }
